% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ras.R
\name{read_ras}
\alias{read_ras}
\title{Read .ras files}
\usage{
read_ras(file, names = c("x", "y"))
}
\arguments{
\item{file}{a path to a ras-file}

\item{names}{the names of the x and y dimension of the data frame as a
character vector. Default values are \code{c("x", "y")}.}
}
\description{
auryn saves its files with a "ras" extension. This function reads in the file
as a data frame.
}
